WIP - Gemming up
Rather than using sudo/rake to install to the users system, install the
binary as part of a rubygem. Doesn't require sudo for users of
RVM/rbenv. Also easier to manage versions.

It's not quite working yet. Also, these commits will likely all be
squashed, so fork this branch at your peril!

Next steps:

Right now, all the work is in bin/old_font - this should be split up
into classes in ./lib which'll make it easier to write tests for.